# td-level-editor/td-level-editor-http/src/main/resources/openapi/td-level-studio-api.yaml
openapi: 3.0.3
info:
  title: Tower Defense Level Studio API
  description: |
    API complète pour l'éditeur de niveaux Tower Defense.
    
    Cette API permet de :
    - Créer et éditer des niveaux
    - Gérer les types d'ennemis
    - Gérer les types de tours
    - Importer/exporter en format draft (éditable) ou exportable (pour le moteur)
  version: 2.0.0
  contact:
    name: Tower Defense Studio Team
servers:
  - url: http://localhost:8081
    description: Serveur de développement local
...
paths:
...
  /api/levels:
    get:
      summary: Liste tous les niveaux en cours d'édition
      tags: [Levels]
      operationId: getAllLevels
      responses:
        '200':
          description: Liste des niveaux
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EditableLevel'
    
    post:
      summary: Crée un nouveau niveau vide
      tags: [Levels]
      operationId: createLevel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLevelRequest'
      responses:
        '201':
          description: Niveau créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditableLevel'
        '400':
          description: Paramètres invalides
  /api/levels/{levelId}:
    get:
      summary: Récupère un niveau par son ID
      tags: [Levels]
      operationId: getLevel
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Détails du niveau
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditableLevel'
        '404':
          description: Niveau non trouvé
    
    delete:
      summary: Supprime un niveau
      tags: [Levels]
      operationId: deleteLevel
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Niveau supprimé avec succès
  /api/levels/import/draft:
    post:
      summary: Importe un niveau depuis un fichier draft (JSON éditable)
      tags: [Levels]
      operationId: importDraftLevel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportRequest'
      responses:
        '201':
          description: Niveau importé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditableLevel'
  /api/levels/import/exportable:
    post:
      summary: Importe un niveau depuis un fichier exportable (format moteur)
      tags: [Levels]
      operationId: importExportableLevel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportRequest'
      responses:
        '201':
          description: Niveau importé avec succès
  /api/levels/{levelId}/export/draft:
    post:
      summary: Exporte un niveau en format draft
      tags: [Levels]
      operationId: exportDraftLevel
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Niveau exporté avec succès
  /api/levels/{levelId}/export/exportable:
    post:
      summary: Exporte un niveau en format exportable (pour le moteur)
      tags: [Levels]
      operationId: exportExportableLevel
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Niveau exporté avec succès

  # ========================================================================
  # LEVEL EDITING
  # ========================================================================
...
  /api/towers:
    get:
      summary: Liste tous les types de tours
      tags: [Towers]
      operationId: getAllTowerTypes
      responses:
        '200':
          description: Liste des types de tours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EditableTowerType'
    
    post:
      summary: Crée un nouveau type de tour
      tags: [Towers]
      operationId: createTowerType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTowerTypeRequest'
      responses:
        '201':
          description: Type de tour créé
  /api/towers/{towerTypeId}:
    get:
      summary: Récupère un type de tour par son ID
      tags: [Towers]
      operationId: getTowerType
      parameters:
        - name: towerTypeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Détails du type de tour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditableTowerType'
    
    delete:
      summary: Supprime un type de tour
      tags: [Towers]
      operationId: deleteTowerType
      parameters:
        - name: towerTypeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Type de tour supprimé
  /api/towers/import/draft:
    post:
      summary: Importe un type de tour depuis un fichier draft
      tags: [Towers]
      operationId: importDraftTowerType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportRequest'
      responses:
        '201':
          description: Type de tour importé
  /api/towers/import/exportable:
    post:
      summary: Importe un type de tour depuis un fichier exportable
      tags: [Towers]
      operationId: importExportableTowerType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportRequest'
      responses:
        '201':
          description: Type de tour importé
  /api/towers/{towerTypeId}/export/draft:
    post:
      summary: Exporte un type de tour en format draft
      tags: [Towers]
      operationId: exportDraftTowerType
      parameters:
        - name: towerTypeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Type de tour exporté
  /api/towers/{towerTypeId}/export/exportable:
    post:
      summary: Exporte un type de tour en format exportable
      tags: [Towers]
      operationId: exportExportableTowerType
      parameters:
        - name: towerTypeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Type de tour exporté

  # ========================================================================
  # TOWER EDITING
  # ========================================================================
  /api/towers/{towerTypeId}/levels:
    post:
      summary: Ajoute un nouveau niveau à un type de tour
      tags: [Tower Editing]
      operationId: addTowerLevel
      parameters:
        - name: towerTypeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditableTowerLevel'
      responses:
        '201':
          description: Niveau de tour ajouté
  /api/towers/{towerTypeId}/levels/{level}:
    delete:
      summary: Supprime un niveau d'un type de tour
      tags: [Tower Editing]
      operationId: removeTowerLevel
      parameters:
        - name: towerTypeId
          in: path
          required: true
          schema:
            type: string
        - name: level
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Niveau de tour supprimé

# ============================================================================
# SCHEMAS
# ============================================================================
components:
  schemas:
...
    CreateLevelRequest:
      type: object
      required:
        - id
        - width
        - height
      properties:
        id:
          type: string
        width:
          type: integer
          minimum: 5
          maximum: 50
        height:
          type: integer
          minimum: 5
          maximum: 50
    ImportRequest:
      type: object
      required:
        - path
      properties:
        path:
          type: string
          description: Chemin du fichier à importer
    ExportRequest:
      type: object
      required:
        - path
      properties:
        path:
          type: string
          description: Chemin où sauvegarder le fichier
...
    CreateTowerTypeRequest:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string

    # ========================================================================
    # DOMAIN MODELS
    # ========================================================================
    EditableLevel:
      type: object
      required:
        - metadata
        - map
        - paths
        - attacks
        - towerCapacityIds
      properties:
        metadata:
          $ref: '#/components/schemas/LevelMetadata'
        map:
          $ref: '#/components/schemas/EditableMap'
        paths:
          type: array
          items:
            $ref: '#/components/schemas/EditablePath'
        attacks:
          type: array
          items:
            $ref: '#/components/schemas/EditableAttack'
        towerCapacityIds:
          type: array
          items:
            type: string
    LevelMetadata:
      type: object
      required:
        - id
        - startingMoney
        - startingLives
      properties:
        id:
          type: string
        startingMoney:
          type: integer
          minimum: 0
        startingLives:
          type: integer
          minimum: 1
    EditableMap:
      type: object
      required:
        - width
        - height
        - paths
      properties:
        width:
          type: integer
        height:
          type: integer
        paths:
          type: array
          items:
            $ref: '#/components/schemas/EditablePath'
    EditablePath:
      type: object
      required:
        - id
        - points
      properties:
        id:
          type: string
        points:
          type: array
          items:
            $ref: '#/components/schemas/PositionDefinition'
          minItems: 2
    PositionDefinition:
      type: object
      required:
        - x
        - y
      properties:
        x:
          type: integer
        y:
          type: integer
    EditableAttack:
      type: object
      required:
        - id
        - waves
      properties:
        id:
          type: string
        waves:
          type: array
          items:
            $ref: '#/components/schemas/EditableWave'
    EditableWave:
      type: object
      required:
        - id
        - startTick
        - spawnInterval
        - count
        - enemyType
        - pathId
      properties:
        id:
          type: string
        startTick:
          type: integer
          minimum: 0
        spawnInterval:
          type: integer
          minimum: 1
        count:
          type: integer
          minimum: 1
        enemyType:
          type: string
        pathId:
          type: string
...
    EditableTowerType:
      type: object
      required:
        - id
        - towerType
        - upgrades
      properties:
        id:
          type: string
        towerType:
          type: string
        upgrades:
          type: array
          items:
            $ref: '#/components/schemas/EditableTowerLevel'
    EditableTowerLevel:
      type: object
      required:
        - level
        - cost
        - range
        - damage
        - reloadTime
        - sellReward
        - buildTimeTicks
      properties:
        level:
          type: integer
        cost:
          type: integer
        range:
          type: number
          format: double
        damage:
          type: integer
        reloadTime:
          type: number
          format: double
        sellReward:
          type: integer
        buildTimeTicks:
          type: integer